Do not use RpcService in srm-impl components
Ditch the dependency on OdlSrmRpcsService and define a more natural
interface, RegistryControl, for use with the shell. This avoids some amount
of boiler plate, as we do not have to deal with the entire RPC dispatch
stack.

While we are here, also clean up pom.xmls to eliminate dependency
warnings.

JIRA: SRVUTILS-10
Change-Id: I218d7b6681270ea72fbc103f15199431a64fe50e
Signed-off-by: lubos-cicut <lubos.cicut@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>